Refactor the original implementation of RMG resonance generation
