--- rdmc/resonance/exceptions.py
@@ -1,5 +1,0 @@
-class ResonanceError(Exception):
-    """
-    An exception class for when RMG is unable to generate resonance structures.
-    """
-    pass
